requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.4
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.5
filelock==3.8.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==23.0
Pillow==9.4.0
platformdirs==2.5.2
pycodestyle==2.9.1
PyJWT==1.7.1
python-dotenv==0.21.1
pytz==2022.4
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
